+++ backend/src/config/route/shipping-address.router.ts
@@ -1,0 +1,6 @@
+import { Express } from 'express';
+import * as shippingAddressCtrl from '@controller/shipping-address.ctrl';
+
+export default function(app: Express, baseURI: string) {
+  app.delete(`${baseURI}/:id`, shippingAddressCtrl.deleteShippingAddress.validator, shippingAddressCtrl.deleteShippingAddress.controller);
+}
