+++ backend/src/service/user.service.ts
@@ -88,70 +88,3 @@ export const updateUserProfile = async (userId: number, nickname: string, imageN
     con.release();
   }
 }
-
-// 해당 사용자의 모든 배송정보 조회
-export const selectUserShippingAddressList = async (userId: number) => {
-  const con = await db.getConnection();
-
-  try {
-    let sql = `
-    SELECT id, user_id, name, recipient, contact, postcode, address, address_detail, requirement
-    FROM ShippingAddress
-    WHERE user_id=${con.escape(userId)}
-    ORDER BY id`;
-
-    interface DataType extends RowDataPacket {
-      id: number;
-      user_id: number;
-      name: string;
-      recipient: string;
-      contact: string;
-      postcode: string;
-      address: string;
-      address_detail: string;
-      requirement: string;
-    }
-
-    return await con.query<DataType[]>(sql);
-  } catch (err) {
-    throw err;
-  } finally {
-    con.release();
-  }
-}
-
-// 사용자 배송정보 추가
-export const insertShippingAddress = async (
-  userId: number,
-  name: string,
-  recipient: string,
-  contact: string,
-  postcode: string,
-  address: string,
-  address_detail: string,
-  requirement: string
-) => {
-  const con = await db.getConnection();
-
-  try {
-    let sql = `
-    INSERT INTO ShippingAddress 
-    (user_id, name, recipient, contact, postcode, address, address_detail, requirement)
-    VALUES (
-      ${con.escape(userId)}, 
-      ${con.escape(name)}, 
-      ${con.escape(recipient)}, 
-      ${con.escape(contact)}, 
-      ${con.escape(postcode)}, 
-      ${con.escape(address)}, 
-      ${con.escape(address_detail)}, 
-      ${con.escape(requirement)}
-    )`;
-
-    return await con.execute(sql);
-  } catch (err) {
-    throw err;
-  } finally {
-    con.release();
-  }
-}
