+++ client/src/page/user/profile/Index.scss
@@ -46,7 +46,10 @@
       flex-wrap: wrap;
 
       .name { margin: 0 0 1em 0; }
-      .text { display: inline-block; margin: 0 0 0.5em 0; }
+      .text {
+        // display: inline-block;
+        margin: 0 0 0.5em 0;
+      }
       .content-section {
         width: 100%;
         display: flex;
